exec function given as parameter on each elemnt of an arr

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	unsigned int long i;
 
 	if (!action)
 		return;
